+++ src/main/java/jpabook/dashdine/service/user/UserInfoService.java
@@ -1,7 +1,7 @@
-package jpabook.dashdine.service;
+package jpabook.dashdine.service.user;
 
 import jpabook.dashdine.domain.user.User;
-import jpabook.dashdine.repository.UserRepository;
+import jpabook.dashdine.repository.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
