+++ src/test/java/jpabook/dashdine/service/user/UserManagementServiceTest.java
@@ -4,9 +4,7 @@ import jpabook.dashdine.domain.user.PasswordManager;
 import jpabook.dashdine.domain.user.User;
 import jpabook.dashdine.domain.user.UserRoleEnum;
 import jpabook.dashdine.dto.request.PasswordChangeRequestDto;
-import jpabook.dashdine.repository.UserRepository;
-import jpabook.dashdine.service.PasswordManagerService;
-import jpabook.dashdine.service.UserManagementService;
+import jpabook.dashdine.repository.user.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
